get pretty 💅

+++ contracts/UniswapFlashQuery.sol
@@ -5,27 +5,41 @@ pragma experimental ABIEncoderV2;
 
 interface IUniswapV2Pair {
     function token0() external view returns (address);
+
     function token1() external view returns (address);
-    function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
+
+    function getReserves()
+        external
+        view
+        returns (
+            uint112 reserve0,
+            uint112 reserve1,
+            uint32 blockTimestampLast
+        );
 }
 
-abstract contract UniswapV2Factory  {
+abstract contract UniswapV2Factory {
     mapping(address => mapping(address => address)) public getPair;
     address[] public allPairs;
-    function allPairsLength() external view virtual returns (uint);
+
+    function allPairsLength() external view virtual returns (uint256);
 }
 
 // In order to quickly load up data from Uniswap-like market, this contract allows easy iteration with a single eth_call
 contract FlashBotsUniswapQuery {
     function getReservesByPairs(IUniswapV2Pair[] calldata _pairs) external view returns (uint256[3][] memory) {
         uint256[3][] memory result = new uint256[3][](_pairs.length);
-        for (uint i = 0; i < _pairs.length; i++) {
+        for (uint256 i = 0; i < _pairs.length; i++) {
             (result[i][0], result[i][1], result[i][2]) = _pairs[i].getReserves();
         }
         return result;
     }
 
-    function getPairsByIndexRange(UniswapV2Factory _uniswapFactory, uint256 _start, uint256 _stop) external view returns (address[3][] memory)  {
+    function getPairsByIndexRange(
+        UniswapV2Factory _uniswapFactory,
+        uint256 _start,
+        uint256 _stop
+    ) external view returns (address[3][] memory) {
         uint256 _allPairsLength = _uniswapFactory.allPairsLength();
         if (_stop > _allPairsLength) {
             _stop = _allPairsLength;
@@ -33,7 +47,7 @@ contract FlashBotsUniswapQuery {
         require(_stop >= _start, "start cannot be higher than stop");
         uint256 _qty = _stop - _start;
         address[3][] memory result = new address[3][](_qty);
-        for (uint i = 0; i < _qty; i++) {
+        for (uint256 i = 0; i < _qty; i++) {
             IUniswapV2Pair _uniswapPair = IUniswapV2Pair(_uniswapFactory.allPairs(_start + i));
             result[i][0] = _uniswapPair.token0();
             result[i][1] = _uniswapPair.token1();
